+++ database/migrations/20191004192833-create-checkout.js
@@ -20,7 +20,7 @@ module.exports = {
         unique: true
       },
       amount: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DOUBLE
       },
       created_at: {
         allowNull: false,
